swiftrev_v3/apps/ai-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
scikit-learn==1.4.0
prophet==1.1.5
pandas==2.2.0
numpy==1.26.3
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.1
httpx==0.26.0
redis==5.0.1
bullmq-python==0.1.0  # Note: Investigating python bullmq alternative if needed
